requirements.txt
Flask
gunicorn
scikit-learn
xgboost
pandas
pyarrow
pytest
category_encoders
requests
google-cloud-storage